@@ -1,6 +1,5 @@
 package ch.akop.homesystem.external.akop;
 
-import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Data;
 
@@ -8,7 +7,5 @@ import lombok.Data;
 public class Heartbeat {
 
   private UUID id;
-  private String gitBranch;
-  private String gitCommit;
-  private LocalDateTime gitCommitDate;
+  private String version;
 }
