@@ -6,21 +6,20 @@ import ch.akop.homesystem.models.devices.actor.DimmableLight;
 import ch.akop.homesystem.persistence.conveter.ColorConverter;
 import ch.akop.homesystem.persistence.model.animation.Animation;
 import ch.akop.homesystem.services.impl.DeviceService;
+import jakarta.persistence.Column;
+import jakarta.persistence.Convert;
+import jakarta.persistence.Entity;
+import jakarta.persistence.EntityNotFoundException;
+import jakarta.persistence.GeneratedValue;
+import jakarta.persistence.Id;
+import jakarta.persistence.ManyToOne;
+import jakarta.persistence.Table;
 import java.math.BigDecimal;
 import java.time.Duration;
 import java.util.UUID;
-import javax.persistence.Column;
-import javax.persistence.Convert;
-import javax.persistence.Entity;
-import javax.persistence.EntityNotFoundException;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import javax.persistence.ManyToOne;
-import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
-import org.hibernate.annotations.GenericGenerator;
 import org.jetbrains.annotations.Nullable;
 
 @Entity
@@ -32,8 +31,7 @@ public class DimmLightStep implements Step {
   public static final Duration DEFAULT_DURATION = Duration.ofSeconds(10);
 
   @Id
-  @GeneratedValue(generator = "UUID")
-  @GenericGenerator(name = "UUID", strategy = "org.hibernate.id.UUIDGenerator")
+  @GeneratedValue
   private UUID id;
 
   @NonNull
