@@ -1,14 +1,13 @@
 package ch.akop.homesystem.persistence.model.config;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
-import org.hibernate.annotations.UpdateTimestamp;
 
 @Getter
 @Setter
@@ -17,7 +16,6 @@ import org.hibernate.annotations.UpdateTimestamp;
 public class DeconzConfig {
 
   @Id
-  @UpdateTimestamp
   private LocalDateTime modified = LocalDateTime.now();
 
   @NonNull
