@@ -1,14 +1,13 @@
 package ch.akop.homesystem.persistence.model.config;
 
 import ch.akop.homesystem.external.openai.ImageRequest;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
-import org.hibernate.annotations.UpdateTimestamp;
 
 
 @Entity
@@ -18,7 +17,6 @@ import org.hibernate.annotations.UpdateTimestamp;
 public class OpenAIConfig {
 
   @Id
-  @UpdateTimestamp
   private LocalDateTime modified = LocalDateTime.now();
 
   @NonNull
