@@ -1,13 +1,12 @@
 package ch.akop.homesystem.persistence.model.config;
 
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
 import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.annotations.UpdateTimestamp;
 
 @Entity
 @Table(name = "config_telegram")
@@ -16,7 +15,6 @@ import org.hibernate.annotations.UpdateTimestamp;
 public class TelegramConfig {
 
   @Id
-  @UpdateTimestamp
   private LocalDateTime modified = LocalDateTime.now();
 
   @Column(columnDefinition = "TEXT")
