@@ -9,6 +9,7 @@ import ch.akop.homesystem.persistence.model.config.TimerConfig;
 import ch.akop.homesystem.persistence.repository.config.TimerConfigRepository;
 import io.reactivex.rxjava3.core.Observable;
 import io.reactivex.rxjava3.disposables.Disposable;
+import jakarta.transaction.Transactional;
 import java.time.Duration;
 import java.time.LocalTime;
 import java.time.ZoneId;
@@ -20,7 +21,6 @@ import java.util.Map;
 import java.util.Set;
 import java.util.function.Consumer;
 import java.util.stream.Stream;
-import javax.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
