@@ -1,8 +1,8 @@
 package ch.akop.homesystem.telemetry;
 
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
 import java.util.UUID;
-import javax.persistence.Entity;
-import javax.persistence.Id;
 import lombok.Getter;
 import lombok.Setter;
 
